+++ config/config.go
@@ -17,13 +17,16 @@ const (
 )
 
 type Config struct {
-	Address           string
-	MetricsAddress    string
-	AvailabilityZones string
-	CertFile          string
-	Logger            micrologger.Logger
-	K8sClient         k8sclient.Interface
-	KeyFile           string
+	Address                  string
+	MetricsAddress           string
+	AvailabilityZones        string
+	CertFile                 string
+	DockerCIDR               string
+	IPAMNetworkCIDR          string
+	KubernetesClusterIPRange string
+	Logger                   micrologger.Logger
+	K8sClient                k8sclient.Interface
+	KeyFile                  string
 }
 
 func Parse() (Config, error) {
@@ -66,8 +69,11 @@ func Parse() (Config, error) {
 	}
 
 	kingpin.Flag("address", "The address to listen on").Default(defaultAddress).StringVar(&config.Address)
+	kingpin.Flag("availability-zones", "List of AWS availability zones").Required().StringVar(&config.AvailabilityZones)
+	kingpin.Flag("docker-cidr", "Default CIDR from Docker").Required().StringVar(&config.DockerCIDR)
+	kingpin.Flag("ipam-network-cidr", "Default CIDR from tenant cluster").Required().StringVar(&config.IPAMNetworkCIDR)
+	kingpin.Flag("kubernetes-cluster-ip-range", "Default CIDR from Kubernetes").Required().StringVar(&config.KubernetesClusterIPRange)
 	kingpin.Flag("metrics-address", "The metrics address for Prometheus").Default(defaultMetricsAddress).StringVar(&config.MetricsAddress)
-	kingpin.Flag("availability-zones", "List of AWS availability zones.").Required().StringVar(&config.AvailabilityZones)
 	kingpin.Flag("tls-cert-file", "File containing the certificate for HTTPS").Required().StringVar(&config.CertFile)
 	kingpin.Flag("tls-key-file", "File containing the private key for HTTPS").Required().StringVar(&config.KeyFile)